+++ SpatialGDK/Source/SpatialGDK/Private/Utils/InterestFactory.cpp
@@ -9,7 +9,6 @@
 
 #include "EngineClasses/Components/ActorInterestComponent.h"
 #include "EngineClasses/SpatialNetConnection.h"
-#include "EngineClasses/SpatialNetDriver.h"
 #include "EngineClasses/SpatialPackageMapClient.h"
 #include "SpatialGDKSettings.h"
 #include "SpatialConstants.h"
@@ -94,11 +93,11 @@ void GatherClientInterestDistances()
 	}
 }
 
-InterestFactory::InterestFactory(AActor* InActor, const FClassInfo& InInfo, USpatialNetDriver* InNetDriver)
+InterestFactory::InterestFactory(AActor* InActor, const FClassInfo& InInfo, USpatialClassInfoManager* InClassInfoManager, USpatialPackageMapClient* InPackageMap)
 	: Actor(InActor)
 	, Info(InInfo)
-	, NetDriver(InNetDriver)
-	, PackageMap(InNetDriver->PackageMap)
+	, ClassInfoManager(InClassInfoManager)
+	, PackageMap(InPackageMap)
 {
 }
 
@@ -257,13 +256,13 @@ Interest InterestFactory::CreatePlayerOwnedActorInterest() const
 void InterestFactory::AddUserDefinedQueries(const QueryConstraint& LevelConstraints, TArray<SpatialGDK::Query>& OutQueries) const
 {
 	check(Actor);
-	check(NetDriver != nullptr && NetDriver->ClassInfoManager);
+	check(ClassInfoManager);
 
 	TArray<UActorInterestComponent*> ActorInterestComponents;
 	Actor->GetComponents<UActorInterestComponent>(ActorInterestComponents);
 	if (ActorInterestComponents.Num() == 1)
 	{
-		ActorInterestComponents[0]->CreateQueries(*NetDriver->ClassInfoManager, LevelConstraints, OutQueries);
+		ActorInterestComponents[0]->CreateQueries(*ClassInfoManager, LevelConstraints, OutQueries);
 	}
 	else if (ActorInterestComponents.Num() > 1)
 	{
@@ -429,8 +428,8 @@ void InterestFactory::AddObjectToConstraint(UObjectPropertyBase* Property, uint8
 
 void InterestFactory::AddTypeHierarchyToConstraint(const UClass& BaseType, QueryConstraint& OutConstraint) const
 {
-	check(NetDriver && NetDriver->ClassInfoManager);
-	TArray<Worker_ComponentId> ComponentIds = NetDriver->ClassInfoManager->GetComponentIdsForClassHierarchy(BaseType);
+	check(ClassInfoManager);
+	TArray<Worker_ComponentId> ComponentIds = ClassInfoManager->GetComponentIdsForClassHierarchy(BaseType);
 	for (Worker_ComponentId ComponentId : ComponentIds)
 	{
 		QueryConstraint ComponentTypeConstraint;
@@ -457,7 +456,7 @@ QueryConstraint InterestFactory::CreateLevelConstraints() const
 	// Create component constraints for every loaded sublevel
 	for (const auto& LevelPath : LoadedLevels)
 	{
-		const uint32 ComponentId = NetDriver->ClassInfoManager->GetComponentIdFromLevelPath(LevelPath.ToString());
+		const uint32 ComponentId = ClassInfoManager->GetComponentIdFromLevelPath(LevelPath.ToString());
 		if (ComponentId != SpatialConstants::INVALID_COMPONENT_ID)
 		{
 			QueryConstraint SpecificLevelConstraint;